chore: 移除不需要的依赖
