+++ package.json
@@ -47,10 +47,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "semver-compare": "^1.0.0",
-    "slate": "^0.63.0",
-    "slate-history": "^0.62.0",
-    "slate-hyperscript": "^0.62.0",
-    "slate-react": "^0.65.0",
     "turndown": "^7.0.0"
   },
   "devDependencies": {
