--- src/components/editor/Toolbar.module.less
@@ -1,53 +1,0 @@
-@import '~@/styles/parameters.less';
-
-.toolbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: @toolbar-background-color;
-  padding: @base-padding;
-  border: 1px solid @toolbar-border-color;
-  border-radius: @toolbar-border-radius;
-  box-shadow: @toolbar-box-shadow;
-  margin-bottom: @toolbar-margin-bottom;
-}
-
-.menu {
-  display: flex;
-  align-items: center;
-
-  & > * {
-    display: inline-block;
-  }
-
-  & > * + * {
-    margin-left: @menu-item-spacing;
-  }
-}
-
-.button {
-  cursor: pointer;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.button.active {
-  color: @black;
-}
-
-.button.inactive {
-  color: @grey-6;
-}
-
-.button.reversed.active {
-  color: @white;
-}
-
-.button.reversed.inactive {
-  color: @grey-7;
-}
-
-.continueBtn {
-  float: right;
-}
