--- src/components/editor/constant.ts
@@ -1,2 +1,0 @@
-export const LIST_TYPES = [ 'numbered-list', 'bulleted-list' ];
-export const TEXT_ALIGN_TYPES = [ 'left', 'center', 'right', 'justify' ];
