+++ src/pages/sandbox/SaveTo.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useContext } from 'react';
+import React, { useState, useEffect, useContext, useRef } from 'react';
 import { ConfigProvider, Button, Select, message, Menu } from 'antd';
 import classnames from 'classnames';
 import { get as safeGet } from 'lodash';
@@ -6,14 +6,9 @@ import type { MenuInfo } from 'rc-menu/lib/interface';
 import Icon, { LinkOutlined } from '@ant-design/icons';
 import Chrome from '@/core/chrome';
 import proxy from '@/core/proxy';
-import processHtmls from '@/core/html-parser';
+import processHTMLs from '@/core/html-parser';
 import LinkHelper from '@/core/link-helper';
-import Editor from '@/components/editor/Editor';
-import serialize from '@/components/editor/serialize';
-import deserialize from '@/components/editor/deserialize';
-import formatHTML from '@/components/editor/format-html';
-import formatMD from '@/components/editor/format-md';
-import contentParser from '@/components/editor/content-parser';
+import LakeEditor, { IEditorRef } from '@/components/lake-editor/editor';
 import { GLOBAL_EVENTS } from '@/events';
 import { EditorValueContext } from './EditorValueContext';
 
@@ -31,48 +26,8 @@ interface RequestMessage {
   htmls?: string[];
 }
 
-let editorInstance;
-
-const getBookmarkContent = (tab: chrome.tabs.Tab) => {
-  return [
-    {
-      type: 'heading-two',
-      reference: true,
-      children: [{ text: tab.title }],
-    },
-    {
-      type: 'paragraph',
-      children: [
-        {
-          type: 'bookmark-link',
-          url: tab.url,
-          children: [{ text: tab.title }],
-        },
-      ],
-    },
-  ];
-};
-
-const getCitation = (tab: chrome.tabs.Tab) => {
-  return [
-    {
-      type: 'paragraph',
-      children: [
-        {
-          text: '来自: ',
-        },
-        {
-          type: 'link',
-          url: tab.url,
-          children: [
-            {
-              text: `${tab.url}`,
-            },
-          ],
-        },
-      ],
-    },
-  ];
+const getBookmarkHtml = (tab: chrome.tabs.Tab) => {
+  return `<h2>${tab.title}</h2><p><a href="${tab.url}">${tab.title}</a></p>`;
 };
 
 const getCurrentTab = (): Promise<chrome.tabs.Tab> =>
@@ -87,21 +42,6 @@ const getNoteId = async (): Promise<string> => {
   const noteId = safeGet(noteStatusResponse, 'data.meta.mirror.id');
   return noteId;
 };
-
-const getPageHTML = (): Promise<string> =>
-  new Promise(resolve => {
-    getCurrentTab().then(tab => {
-      Chrome.tabs.sendMessage(
-        tab.id,
-        {
-          action: GLOBAL_EVENTS.GET_PAGE_HTML,
-        },
-        html => {
-          resolve(html);
-        },
-      );
-    });
-  });
 
 const NOTE_DATA = {
   id: 0,
@@ -143,15 +83,16 @@ function BookWithIcon({ book }) {
 const useViewModel = props => {
   const [ books, setBooks ] = useState([ NOTE_DATA ]);
   const [ currentBookId, setCurrentBookId ] = useState(NOTE_DATA.id);
-  const { editorValue, currentType, setEditorValue, setCurrentType } =
+  const editorRef = useRef<IEditorRef>(null);
+  const { currentType, setCurrentType } =
     useContext(EditorValueContext);
   const onSelectType = setCurrentType;
 
   const startSelect = (append = false) => {
     getCurrentTab().then(tab => {
-      const bookmarkContent = getBookmarkContent(tab);
       if (!append) {
-        setEditorValue(bookmarkContent);
+        const html = getBookmarkHtml(tab);
+        editorRef.current?.setContent(html);
       }
       Chrome.tabs.sendMessage(tab.id, {
         action: GLOBAL_EVENTS.START_SELECT,
@@ -178,27 +119,22 @@ const useViewModel = props => {
       case GLOBAL_EVENTS.GET_SELECTED_HTML: {
         const { htmls } = request;
         const noteId = await getNoteId();
-        const processedHtmls = await processHtmls(htmls, noteId);
-        const newHtml = processedHtmls.map(html => formatHTML(html)).join('');
-        const document = new window.DOMParser().parseFromString(
-          newHtml,
-          'text/html',
-        );
-        const value = deserialize(document.body);
-        setEditorValue([ ...editorValue, ...formatMD(value) ]);
+        const processedHTMLs = await processHTMLs(htmls, noteId);
+        if (editorRef.current.isEmpty()) {
+          const initHtml = getBookmarkHtml(await getCurrentTab());
+          editorRef.current?.appendContent(initHtml);
+        }
+        editorRef.current?.appendContent(processedHTMLs.join(''), true);
         sendResponse(true);
         return;
       }
       case GLOBAL_EVENTS.GET_SELECTED_TEXT: {
         const { htmls } = request;
-        const newHtml = htmls.map(html => formatHTML(html)).join('');
-        const document = new window.DOMParser().parseFromString(
-          newHtml,
-          'text/html',
-        );
-
-        const value = deserialize(document.body);
-        setEditorValue([ ...editorValue, ...formatMD(value) ]);
+        if (editorRef.current.isEmpty()) {
+          const initHtml = getBookmarkHtml(await getCurrentTab());
+          editorRef.current?.appendContent(initHtml);
+        }
+        editorRef.current?.appendContent(htmls.join(''), true);
         setCurrentType('selection');
         sendResponse(true);
         return;
@@ -213,31 +149,15 @@ const useViewModel = props => {
     return () => {
       Chrome.runtime.onMessage.removeListener(onReceiveMessage);
     };
-  }, [ editorValue ]);
+  }, []);
 
   useEffect(() => {
     if (currentType === SELECT_TYPES[0].key) {
       startSelect();
     } else if (currentType === SELECT_TYPES[1].key) {
       getCurrentTab().then(tab => {
-        const bookmarkContent = getBookmarkContent(tab);
-        setEditorValue(bookmarkContent);
-      });
-    } else if (currentType === SELECT_TYPES[2]?.key) {
-      getPageHTML().then(res => {
-        const html = formatHTML(res);
-        const document = new window.DOMParser().parseFromString(
-          html,
-          'text/html',
-        );
-        const value = deserialize(document.body);
-        const formattedValue = formatMD(value);
-        setEditorValue(editorValue.concat(formattedValue));
-      });
-    } else if (currentType === 'selection') {
-      getCurrentTab().then(tab => {
-        const citation = getCitation(tab);
-        setEditorValue(editorValue.concat(citation));
+        const html = getBookmarkHtml(tab);
+        editorRef.current?.setContent(html);
       });
     }
   }, [ currentType ]);
@@ -245,23 +165,12 @@ const useViewModel = props => {
   const [ loading, setLoading ] = React.useState<boolean>(false);
 
   const onSave = () => {
-    if (!editorInstance) return;
-    const shouldAddReferenceNode = currentBookId === NOTE_DATA.id;
+    if (!editorRef.current) return;
 
-    const serializedAsiContent = contentParser.wrapLakeAslBody(
-      editorInstance.children
-        .map(node => serialize(node, true, shouldAddReferenceNode))
-        .join(''),
-    );
+    const serializedAsiContent = editorRef.current?.getContent('lake') || '';
+    const serializedHtmlContent = editorRef.current?.getContent('text/html') || '';
 
-    const serializedHtmlContent = contentParser.wrapLakeHtmlBody(
-      editorInstance.children
-        .map(node => serialize(node, false, shouldAddReferenceNode))
-        .join(''),
-    );
-
-    const onSuccess = (type: 'doc' | 'note', noteOrDoc: any) => {
-      setEditorValue([]);
+    const onSuccess = (type: 'doc' | 'note', noteOrDoc: { id: string }) => {
       setCurrentType(null);
       setLoading(false);
 
@@ -340,10 +249,10 @@ const useViewModel = props => {
   return {
     state: {
       books,
-      editorValue,
       currentBookId,
       currentType,
       loading,
+      editorRef,
     },
     onSave,
     onContinue,
@@ -352,10 +261,10 @@ const useViewModel = props => {
   };
 };
 
-const SaveTo: React.FC = props => {
-  const { currentType, editorValue } = useContext(EditorValueContext);
+const SaveTo = React.forwardRef<IEditorRef, any>((props, ref) => {
+  const { currentType } = useContext(EditorValueContext);
   const {
-    state: { books, currentBookId, loading },
+    state: { books, currentBookId, loading, editorRef },
     onSelectBookId,
     onSave,
     onContinue,
@@ -415,17 +324,19 @@ const SaveTo: React.FC = props => {
           {currentBookId === NOTE_DATA.id ? __i18n('小记') : __i18n('知识库')}
         </Button>
         {currentType && (
-          <div className={styles.editor}>
-            <Editor
-              onLoad={editor => (editorInstance = editor)}
-              defaultValue={editorValue}
-              onClipContinueClick={onContinue}
-            />
+          <div className={styles['lake-editor']}>
+            <LakeEditor ref={editorRef} value="" onChange={html => {
+              // console.info(html);
+            }}>
+              <Button onClick={onContinue}>
+                {__i18n('继续选取')}
+              </Button>
+            </LakeEditor>
           </div>
         )}
       </div>
     </ConfigProvider>
   );
-};
+});
 
 export default SaveTo;
